pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wilds-bridge"
version = "0.1.0"
description = "LDT ActiveMQ bridge for the WILDS spectrograph"
requires-python = ">=3.10"
dependencies = [
    "stomp.py>=4.1.21",
    "pydantic>=2.0",
    "pydantic-xml>=2.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/wilds"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade (enforces modern Python idioms incl. PEP 585/604)
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long — formatter handles this
]

[tool.ruff.lint.isort]
known-first-party = ["wilds"]
